requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.1
SQLAlchemy==1.4.35
Werkzeug==2.1.1
Flask-Login==0.6.0
python-dotenv==0.15.0
abc==2.0
bcrypt==3.2.0
Cerberus==1.3.4
psycopg2==2.8.6
